gps/requirements.txt
-i https://pypi.org/simple
pyserial
RPi.GPIO
pynmea2
geojson
